png work, fixed streaming gg me

+++ source/modules/graphics/wrap_graphics.h
@@ -18,7 +18,7 @@ void graphicsPoints(float x, float y);
 
 void graphicsCircle(float x, float y, float radius, float segments);
 
-void graphicsDraw(C3D_Tex * texture, float x, float y, float width, float height);
+void graphicsDraw(C3D_Tex * texture, float x, float y, int width, int height);
 
 void graphicsSetColor(int r, int g, int b);
